+++ src/sqlalchemy_cratedb/dialect.py
@@ -21,6 +21,7 @@
 
 import logging
 from datetime import date, datetime
+from types import ModuleType
 
 from sqlalchemy import types as sqltypes
 from sqlalchemy.engine import default, reflection
@@ -212,6 +213,12 @@ class CrateDialect(default.DefaultDialect):
         # get default schema name
         self.default_schema_name = self._get_default_schema_name(connection)
 
+    def set_isolation_level(self, dbapi_connection, level):
+        """
+        For CrateDB, this is implemented as a noop.
+        """
+        pass
+
     def do_rollback(self, connection):
         # if any exception is raised by the dbapi, sqlalchemy by default
         # attempts to do a rollback crate doesn't support rollbacks.
@@ -230,7 +237,21 @@ class CrateDialect(default.DefaultDialect):
             use_ssl = asbool(kwargs.pop("ssl", False))
             if use_ssl:
                 servers = ["https://" + server for server in servers]
-            return self.dbapi.connect(servers=servers, **kwargs)
+
+            is_module = isinstance(self.dbapi, ModuleType)
+            if is_module:
+                driver_name = self.dbapi.__name__
+            else:
+                driver_name = self.dbapi.__class__.__name__
+            if driver_name == "crate.client":
+                if "database" in kwargs:
+                    del kwargs["database"]
+                return self.dbapi.connect(servers=servers, **kwargs)
+            elif driver_name in ["psycopg", "PsycopgAdaptDBAPI", "AsyncAdapt_asyncpg_dbapi"]:
+                return self.dbapi.connect(host=host, port=port, **kwargs)
+            else:
+                raise ValueError(f"Unknown driver variant: {driver_name}")
+
         return self.dbapi.connect(**kwargs)
 
     def do_execute(self, cursor, statement, parameters, context=None):
@@ -300,10 +321,12 @@ class CrateDialect(default.DefaultDialect):
         if schema is None:
             schema = self._get_effective_schema_name(connection)
         cursor = connection.exec_driver_sql(
-            "SELECT table_name FROM information_schema.tables "
-            "WHERE {0} = ? "
-            "AND table_type = 'BASE TABLE' "
-            "ORDER BY table_name ASC, {0} ASC".format(self.schema_column),
+            self._format_query(
+                "SELECT table_name FROM information_schema.tables "
+                "WHERE {0} = ? "
+                "AND table_type = 'BASE TABLE' "
+                "ORDER BY table_name ASC, {0} ASC"
+            ).format(self.schema_column),
             (schema or self.default_schema_name,),
         )
         return [row[0] for row in cursor.fetchall()]
@@ -326,7 +349,7 @@ class CrateDialect(default.DefaultDialect):
             "AND column_name !~ ?".format(self.schema_column)
         )
         cursor = connection.exec_driver_sql(
-            query,
+            self._format_query(query),
             (
                 table_name,
                 schema or self.default_schema_name,
@@ -366,7 +389,9 @@ class CrateDialect(default.DefaultDialect):
                 rows = result.fetchone()
                 return set(rows[0] if rows else [])
 
-        pk_result = engine.exec_driver_sql(query, (table_name, schema or self.default_schema_name))
+        pk_result = engine.exec_driver_sql(
+            self._format_query(query), (table_name, schema or self.default_schema_name)
+        )
         pks = result_fun(pk_result)
         return {"constrained_columns": sorted(pks), "name": "PRIMARY KEY"}
 
@@ -404,6 +429,17 @@ class CrateDialect(default.DefaultDialect):
         """
         server_version_info = self.server_version_info
         return server_version_info is not None and server_version_info >= (4, 1, 0)
+
+    def _format_query(self, query):
+        """
+        When using the PostgreSQL protocol with drivers `psycopg` or `asyncpg`,
+        the paramstyle is not `qmark`, but `pyformat`.
+
+        TODO: Review: Is it legit and sane? Are there alternatives?
+        """
+        if self.paramstyle == "pyformat":
+            query = query.replace("= ?", "= %s").replace("!~ ?", "!~ %s")
+        return query
 
 
 class DateTrunc(functions.GenericFunction):